fix: Opencell 함수 BoardCalculator -> Minesweeper 이동
