+++ src/main/kotlin/mine/domain/BoardCalculator.kt
@@ -4,34 +4,7 @@ import mine.dto.Coordinate
 import mine.enums.MineCell
 
 class BoardCalculator {
-    fun openCells(
-        mineBoard: List<MineRow>,
-        coordinate: Coordinate,
-    ) {
-        val (x, y) = coordinate
-        val cell = mineBoard[x].mineCells[y]
-        if (cell.isOpen) return
-        cell.withOpen()
-        when (cell) {
-            is MineCell.Number -> {
-                selectedCellIfSafety(cell, mineBoard, coordinate)
-            }
 
-            else -> return
-        }
-    }
-
-    private fun selectedCellIfSafety(
-        cell: MineCell.Number,
-        mineBoard: List<MineRow>,
-        coordinate: Coordinate,
-    ) {
-        if (cell.value == CELL_ZERO) {
-            mineBoard.forEachIndexed { rowIndex, row ->
-                UpdateMineRow(row = row, coordinate = coordinate, rowIndex = rowIndex).updateCells()
-            }
-        }
-    }
 
     fun isMineCell(
         mineBoard: List<MineRow>,
